Making RaftReplicationParam part of every hook call
Summary:
This will help us latch more things in this param in the
future, thus having an agile way of passing more things to the plugin.
One of the ideas is to make replication multitenant and for that we need
to carry an index to route to the right replication channel.

Differential Revision: D46570576

fbshipit-source-id: 5af551ce1e7fb913c6249c966b224537b9c4be17