Set rocksdb_large_prefix to ON and deprecate this system variable (#1322)
Summary:
This variable was introduced to mirror innodb_large_prefix, which was deprecated in 5.7 and removed in 8.0, and no longer serves any real purpose, nor there is anything to gain with its 'OFF' setting, which is also incompatible with the data dictionary schema. Thus 1) change the default to 'ON'; 2) deprecate it.

Squash with 749f1dc5ef0327e55616441b399000510db9a54d

Pull Request resolved: https://github.com/facebook/mysql-5.6/pull/1322

Differential Revision: D46773511