Fix allow_binlog_writes_on_raft_follower for primary
Summary:
Today, setting this var allows us to write to binlog directly on
followers. We use this to skip GTIDs if needed like in S342118 for
example.

However, turning this on in a leader causes the leader to silently skip
replication, as we've seen in S342118. Enforce the check here such that
this variable only applies for followers.

Differential Revision: D46778138