Fix ROCKSDB_PLUGIN typo in get_rocksdb_files.sh
I'm guessing this was copy/pasted from a Makefile sort of syntax, whereas it meant to include the contents of these environment variables in this bash script.  When I tried to build mysqld from this it was crashing here saying things like `ROCKSDB_PLUGIN_BUILTINS: command not found` because this syntax in bash means to execute a subcommand.