Fix thread_pool/t/is_columns_is_cs.test regression
Summary:
Due to the split nature of fbcode and the mysql git repos, any time we
add a new column to thread_pool plugin's I_S tables (or reorder, or
change type, etc), and after contbuild produces the next fbpkg for the
plugin, the expected results will cease to match the actual ones.

While the fix here is simple: updating the results to include the
current reality, a bigger question remains around what can be done to
mitigate this loose coupling from a test robustness standpoint.

While we contemplate on that, this diff refreshes the expected results
of the column metadata for the TP I_S tables.

Differential Revision: D48039661