Wait on before_commit on largest opid trx, not last
Summary:
before_commit waits for a transaction to be consensus committed. We
currently wait on the last THD within a group. This works if the
transactions in the group is ordered by their order in the binlog. We've
run Raft this way for most of the current rollouts.

However, this is not safe if the transaction in the group are not
ordered this way, as with some of the cases we have (ensuring order in
those cases would have made the applier too slow).

This is because the last THD in the group may not actually be the latest
transaction. Say it's OpID is 4:100. There could be transactions earlier
in the group that are at say OpID 4:110. If the commit point is 4:105,
the before_commit will return, but we'll then proceed to commit 4:110
before it's consensus committed.

Normally we expect transactions to eventually reach consensus commit, so
we don't see any inconsistency here. However, if a DMP happens at this
point, we could truncate and abort transactions above 4:105. In this
case, since we committed 4:110 to the engine, we would be inconsistent.

The fix here takes the max OpID in the group instead of using the last
one. In the case above, it would have been waiting on 4:110 to clear
before_commit. When the DMP happens, before_commit would fail and we'll
rollback, and 4:110 won't be committed.

Differential Revision: D48076348