Fixing idempotent recovery cleanup
Summary:
In case where the last transaction that runs in idempotent mode
is also part of gtid_executed we did not reset
`THD::m_skip_row_logging_functions` which led to transactions without
row events in the apply logs. Moving some methods around so that the
cleanup is intuitive. The following changes were made:
 - `check_and_set_idempotent_recovery` moved from `Log_event` to
   `Rows_log_event` so that all logic can be consolidated within
   `Rows_log_event::do_apply_event()`
 - Using a scope guard in `Rows_log_event::do_apply_event()` to reset
   the state when we return from the method, so in all cases we end up
   cleaning up the state.
 - Got rid of `Slave_worker::worker_last_gtid`. Since `Slave_worker`
   inherits from `Relay_log_info` we can just use
   `Relay_log_info::last_gtid`. This way MTS and single threaded code
   can be similar.

Differential Revision: D48158758

fbshipit-source-id: eb059ab696a5226c8cbc43da280e217b31d07d16