Avoid big allocations in rapidjson
Summary:
It seems like we're doing big heap allocations (64KB) when parsing json, and this is cpu inefficient probably because we're not able to make use of jemalloc tcache.

Our JSON objects are usually small, so we can just use stack allocated memory instead (with fallback to heap allocated memory). This approach is documented in https://rapidjson.org/md_doc_dom.html#UserBuffer

Differential Revision: D48253817