Clear cached memory on PinnableSlice
Summary:
When fetching keys that are in memtable, the value is memcpy'd to a `std::string` object which then backs the memory used by `PinnableSlice`. The issue is that `std::string` caches its memory, meaning that this memory is held until the handler object is destroyed.

Solution is similar to D43753421. Shrink the cache memory when it exceeds some threshold.

Differential Revision: D48400318