avoid acquiring LOCK_global_sql_findings when there is no error/warning/note
Summary:
today we lock LOCK_global_sql_findings for every query, which is wasteful since we do not need to record anything if the query succeeded with no error/warning/note messages.

behaviour change: today we insert an empty SQL_FINDING_VEC to global_sql_findings_map when the query does not have any messages, now we do not do it.

Squash with D25292665

Differential Revision: D48406547

fbshipit-source-id: d8f09d9230ef07a8e0e86b2bdc104287314f9965