Reimplement InnoDB data dictionary INFORMATION_SCHEMA views through server layer (#1342)
Summary:
Currently, the DD-related INFORMATION_SCHEMA views in InnoDB use the low level InnoDB APIs directly to access the underlying DD tables. That stops working when those tables are moved to a different storage engine. To keep these views working, re-implement them using the server layer engine-agnostic APIs. Leave the old implementation enabled when InnoDB is the data dictionary storage engine.

To test the implementation without a viable alternate DDSE yet, the new implementation was enabled unconditionally and all the test regressions fixed and collected to a new testsuite rocksdb_dd_innodb. Going forward it will collect other InnoDB-specific tests when it is not the DDSE.

Pull Request resolved: https://github.com/facebook/mysql-5.6/pull/1342

Differential Revision: D48700405