Fix for snapshot conflict on SK due to concurrent update
Summary: The issue that was being hit has two different connections to the MySQL server trying to update two different rows via a secondary index scan with RC isolation level set. During the SK index scan, either one of the iterators may end up returning the value updated by the other connection. This leads to the iterator range condition failing, and the iterator returns an early EOF back, and the update fails to apply on the remaining rows that were actually valid candidates for update based on the WHERE conditions.

Differential Revision: D48740767

fbshipit-source-id: f9b58edc249b6deace936dbafc70f79456eb49ca