Add option to ignore killed flag for materialization
Summary:
When queries time out, we abandon any work that was done to materialize partial indexes. This may not be desirable since the client will most likely retry, and still fail because it is querying a large group that requires materialization.

Add a flag for the materialization code path to ignore the kill flag and finish materialization. In case we have bugs, we can always turn it off via a flag to check the kill flag.

Differential Revision: D48786848

fbshipit-source-id: 7a9fd996b4a403392c6005dacbea43a23240f568