Checking for errors when copying binlog cache to raft IO cache
Summary:
We were not checking for error when copying binlog cache to
raft IO cache. This can cause partial transaction with only GTID and
Metadata event to be written to the binlog. Now on error we'll crash.
This is fine since we expect this to happen only on disk full and other
HW issues. Also simplified `MYSQL_BIN_LOG::write_metadata_event` in the
process.

Differential Revision: D48854982