Fix group by loose index scan costing
Summary:
In a previous diff (D28915153), I added support for considering LIMIT clause when doing costing, but it wasn't very effective because this was done only after io cost was already calculated. The io cost is usually the dominant cost, so it only sometimes improved plans.

The existing code for computing io cost was complicated and probably has some bugs (documented in https://bugs.mysql.com/112280) so I simplified it to be more based on counting number of index_read seek operations.

Squash with: D28915153

Differential Revision: D49034596

fbshipit-source-id: 417aedf43a7fcb4beb7f84504a12500e086d5340