Fix column usage info size accounting for max_column_statistics_size
Summary:
In the original change D42811784 Ritwik must have forgotten that `thd->column_usage_info` is not `ColumnUsageInfo` but `std::set<ColumnUsageInfo>`. As a result, the function `ColumnUsageInfo::size()` he added was never called, and the code was simply calling `std::set::size()` and using the number of items in the set as the size. Fix this by iterating over the set, adding all ColumnUsageInfo sizes as well as the `std::set` and `std::unordered_map` overheads.

Squash with: D42811784.

Differential Revision: D49179868