Rework MyRocks transaction DDL flag
class Rdb_transaction had a field, m_ddl_transaction, which was effectively a
cached flag for the current SQL statement belonging to a certain subset of DDL
commands. This was used primarily to check whether key uniqueness on write can
be skipped (and secondarily to skip the "Applier is committing transaction
without binlog position" warning).

As a cached value, it was not always set when it should have been, which is
possible to see by adding asserts at the flag read sites that check the flag
value against the THD SQL statement type: rocksdb.commit_without_prepare,
rocksdb.add_index_inplace_sstfilewriter. While it is possible to add the missing
updates, it is much simpler not to cache it in the first place and just check
the THD SQL statement type directly.

At the same time, as the DD tables are being converted to MyRocks, and system
tables are planned too, many more commands can skip the uniqueness checks.

Thus remove the flag, remove its setters and getters, and introduce a helper
is_safe_for_wb with a compilation-time lookup array.