Tweak "Applier is committing transaction without binlog position info" warning conditions
Besides other conditions, instead of skipping it for CREATE INDEX, DROP INDEX, &
ALTER TABLE, skip it for all commands that may not generate row events. As
CREATE TABLE can generate them in the most general case (as CREATE TABLE ...
SELECT), further skip it for CREATE TABLE that is not of CREATE TABLE ... SELECT
form.