Tweak "Applier is committing transaction without binlog position info" warning conditions (#1368)
Summary:
Besides other conditions, instead of skipping it for CREATE INDEX, DROP INDEX, & ALTER TABLE, skip it for all commands that may not generate row events. As CREATE TABLE can generate them in the most general case (as CREATE TABLE ... SELECT), further skip it for CREATE TABLE that is not of CREATE TABLE ... SELECT form.

Pull Request resolved: https://github.com/facebook/mysql-5.6/pull/1368

Differential Revision: D49478330

fbshipit-source-id: 5636fe3f60958895da9e95f6efa4308672a9f589