Fb mysql 8.0.28 clone plugin autotune (#1369)
Summary:
This is a fix for the overshooting upstream bug.
The active_task is calculated in a wrong way. In the current calculation, the active_task will always be 1 and the task_target will be always equal to target_speed, which will lead up to the case that all threads use their max bandwidth. (No throttling at all).
The fix is to change the active_task to the number of threads, since all threads should be running and active.

Pull Request resolved: https://github.com/facebook/mysql-5.6/pull/1369

Differential Revision: D49505544