Flushing engine logs when switching from primary to secondary
Summary:
We need to make sure that we flush the engine logs (aka sync the engines)
before we switch from binlog to apply log (i.e. switch from primary to
secondary) to recover correctly in host restart case. Otherwise if host
restarts after apply log has been created and the engine goes back in
time on recovery, we'll use the apply log for recovery and assume that
engine is caught up.

Differential Revision: D50385170

fbshipit-source-id: 1a0dff6b2586b36f7ee3e9d2322e403d63acc31a