Fix wait_for_ack mtr test
Summary:
In D49558808, we changed defaults for a bunch of system variables. Specifically, the one that affected our tests was: `rpl_raft_async_notify_commit_index`. In that diff, we flipped the default from ON to OFF.

In the test case, we set a debug point at `ReplicationDoneCallback`. This is meant to block the before_commit() from returning on the leader node. However, since rpl_raft_async_notify_commit_index is also OFF, we are synchronously going through the NotifyCommitIndex path which will eventually block on the debug point. This blocks all further replication because PeerMessageQueue executes under a threadpool with ThreadPool::ExecutionMode::SERIAL thread pool. So if we block on the debug point, all replication stops.

With `rpl_raft_async_notify_commit_index = ON`, we are safe to replicate still because we don't block the lone thread and replication continues.

Essentially, this is just a testing artifact. It does not have any implications in production.

Differential Revision: D51216368