mtr: Fix perfschema.statement_program_lost_inst in ASAN
Summary:
This test explicitly overrides the thread stack size to 640K, which has
the effect of shrinking it from the 10MB default in ASAN mode. Even in
non-ASAN, it still shrinks it dramatically from the 1MB default set in
https://github.com/facebook/mysql-5.6/commit/907e8ab52b861fb19e9b9b185970bf10bdbfc65f

This results in failures like:
```
2023-11-11T03:20:50.064432Z 0 [Note] [MY-010902] [Server] Thread priority attribute setting in Resource Group SQL shall be ignored due to
unsupported platform or insufficient privilege.
mysqld: Thread stack overrun:  624144 bytes used of a 655360 byte stack, and 81920 bytes needed.  Use 'mysqld --thread_stack=#' to specify a bigger stack.
2023-11-11T03:20:50.065493Z 0 [ERROR] [MY-010944] [Server] Resource group post initialization failed.
2023-11-11T03:20:50.067230Z 0 [ERROR] [MY-010119] [Server] Aborting
```

It was once an increase from the default 280K size though. But that has
not been the case for years.

I'm removing the override and allowing the default to take care of
itself.

Note: the changes in the .result are all whitespace, due to trimming of
trailing spaces in my editor.

Differential Revision: D51232613

fbshipit-source-id: b0c777e0fd3749ce5bbfe1fc423d9fc530ef827c