Remove Rdb_transaction::m_ddl_transaction
With DD tables in MyRocks, the property of whether it is safe to skip uniqueness
checks on writes should be resolved on a table, not transaction level. Thus fold
the logic of setting m_ddl_transaction to true into ha_rocksdb::skip_unique. At
the same time this flag is duplicated between struct update_row_info and
callstack of affected functions, thus remove it from the callstack.