Add sys var to skip plugin registration by INSTALL PLUGIN
Summary: Currently MySQL lacks the ability to load plugins at runtime without permanent registration. This forces us to restart instances to achieve this one time load with --plugin-load parameter. This change adds the same capability to the `install plugin` command by adding the sys var. We'll set it to `true` for our production so that anyone running `install plugin` will not be able to get the mysql prod instances into the permanent plugin load state that is hard to recover from.

Differential Revision: D51722313

fbshipit-source-id: 67aa9171d1d2075d5dd8d3ae7bc66c86a8361d44