Remove opt_commit_consensus_wait_max_marker
Summary:
This was from a fix for a correctness problem. Because we need to wait
for the max opid in the group commit, not just the last one, if we have
out of order commits.

The flag was there in case we see anything unexpected, but we have not.
Clean it up.

Differential Revision: D53359652

fbshipit-source-id: ee2bd14ebe966aa54163ffa6c88f3078d126dfe6