Allow ignoring thread priority settings
Summary: This diff allows the ignoring of thread priority settings. The changes include adding a new function to the thread_pool_priv.h header file and a new variable to the sql_class.h header file. The sql_class.cc file includes the implementation of the new function and the variable. The changes are made to allow for the ignoring of thread priority settings when we are managing resources via thread pool.

Differential Revision: D53679574

fbshipit-source-id: 28544f070a737db282bf2451009e797b9c83a384