Avoid setting binlog.write_error on cache write errors
Summary:
Currently, if we hit any cache write error, ie, creating a trx that exceeds max_binlog_cache_size, we will set the binlog's write_error var to true. Subsequent transactions will not be able to write to the binlog anymore as we do not allow commits if there are write errors.

The problem is that anytime we hit a cache write error, we have to change the leader. For customers with large write transactions like the deletion framework, this would mean constant dead master promotions.

To avoid this, the fix is to avoid setting write_error on cache write errors. When there is a cache write error, we already reset the cache, so it's safe to resume writes.

Differential Revision: D53682383

fbshipit-source-id: ccc8c1da083c4f92e9fd1e198dc8dbdfa5e7f300