Add additional CPU scheduler APIs
Summary:
Introduce a few APIs to cover the scenarios for thrift plugin and child mysql tasks.
 - `tp_create_connection` takes additional parameter whether or not to acquire connection slot against `thread_pool_max_db_connections` and the resource shape quota. The child tasks would set this parameter to `false`.
 - `tp_is_scheduler_enabled` helps determine if the scheduler is enabled or not. If not, the code should fall back to the other APIs (mysql or thrift).
 - `tp_get_current_task_connection` returns connection of the current task, from which the tenant id could be obtained and used to enqueue child tasks.
 - `tp_get_tenant_id` looks up tenant id by db name and allows enqueueing a task without `THD` or `TpConn` which would need to be created/found on the task itself, and then attached.

Another small change is to get rid of the assert in `THD::get_connection_attr`. The scheduler could call this API when it is about to enqueue a new task for a `THD` which is done from a random thread. The `THD` in question doesn't have any current thread at that moment so it is safe to call this API and access the memory.

Differential Revision: D53980947

fbshipit-source-id: 2ef3d6f79a9375413e252b6070827d129b42848e