Skip checking other waiters after granting MDL mutex
Summary: The `MDL_mutex` is used in place of the regular mutex where a thread could wait for a long time while holding the mutex. In particular, this is used in `Global_THD_manager` to protect the list of `THD`s during `THD` deletions. Under the extreme stress of new connections the deletions slow down due to the large convoy forming on `MDL_mutex`. I noticed that the `MDL_lock` code is generic and tries to check every waiter in all cases when it grants the locks. However, in the MDL mutex case this is not needed, and after the first waiter is granted, all others will be conflicting. So there's no point in traversing a large list of waiters. This fix should help speed up the deletions of `THD`s under heavy connection stress.

Differential Revision: D55439597

fbshipit-source-id: 75048d652659871c7c7cd7bf0c542fa0b7820a30