mysql: add dump column list to DUMP TABLE stmt
Summary:
This allows syntax like

```
DUMP TABLE <table> [ ( <col1> [, <col2>, ... ] ) ]
...
```
E.g.
```
DUMP TABLE t1 INTO 'foo'; -- as before
DUMP TABLE t1 (a, b, c) INTO 'foo'; -- new
```
to allow for a finer grain of selection of columns to dump.

The motivation is to support OSC's usecase which dumps only those
columns needed for the destination table as an optimization. I.e. if
a column is dropped in the after-schema, we don't need to dump it.

Differential Revision: D55775540

fbshipit-source-id: 60d6abd8a0070daeac0d72df30705fba54560108