Release jobs_lock before signalling the dependency replication worker threads
Summary:
Under certain circumstances, the worker's jobs_lock can be acquired twice by the
coordinator thread trying to signal the worker threads.

If the worker was in retry_transaction() where worker_sleep() acquires
the jobs_lock, THD::awake() will also acquire the jobs_lock in order to
signal the conditional variable which the worker thread is waiting on.

If the coordinator thread holds the worker's jobs_lock while calling
THD::awake(), it will attempt to acquire the jobs_lock again and
deadlock.

Fix is to release the jobs_lock before calling awake.

Squash with D19036547

Differential Revision: D55651636

fbshipit-source-id: 291be7eedcc67aed2d3d60a79f75c598db7a7c59