Cleanup rdb_hexdump (#1442)
Summary:
Currently, `rdb_hexdump` has a third argument `maxsize` with a default value of 0,
meaning unlimited output length, and there is also a `RDB_MAX_HEXDUMP` constant
for the callers to use. However, only a part of the callers use it, the rest use
the default or something else altogether (i.e. `FN_REFLEN`)

Change as follows:
- Default to `RDB_MAX_HEXDUMP` `maxsize` instead of unlimited. This should cover
  all the reasonable cases, and prevent runaway dumping in the case of corrupted
  `data_len` values.
- Bump `RDB_MAX_HEXDUMP_LEN` from 1000 to 7000. This way it can dump any MyRocks
  key (3072 byte maximum). Make it `constexpr`.

Minor cleanups:
- Where `std::string` is used as an input for `rdb_hexdump`, pass its buffer
  through `data()`, not through `c_str()`, because it's used as a sized buffer
  and not as a zero-terminated C string
- Mark `rdb_hexdump` as `nodiscard`, make its constants `constexpr`, remove
  redundant `const` from by-value parameters.

Pull Request resolved: https://github.com/facebook/mysql-5.6/pull/1442

Differential Revision: D55594025

fbshipit-source-id: d048078a41c537c4ba4dbdbf1e80c07b33e7f39f